move user types to own file
